+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2] - 2026-07-08
+- Fixed pruning of title fields in byteit schemas
+- Improved error handling for downloading result
 
 ## [1.1.1] - 2026-06-05